requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
psycopg2==2.8.6
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
